requirements.txt
google-generativeai
transformers
torch
python-dotenv
textblob
flask
